requirements.txt

streamlit>=1.34.0
opencv-python-headless>=4.9.0
face_recognition>=1.3.0
scikit-learn>=1.3.0
numpy>=1.26.0
Pillow>=10.3.0
moviepy>=1.0.3
